02/09/2023, Updated Files - Tindog Progress
I made a small amount of progress on a Tinder clone today (and my knowledge and experience with bootstrap as well). I wish I had more time to dedicate to coding today, but, this will have to do. I'm happy that my new schedule should allocate a lot more time for it!