02/14/2023, Updated Files - Code Refactoring, Combining Selectors, and Selector Priority
I learned A TON today nearing the end of the bootstrap module--and I finished my Tinder clone! Everything I learned can be found in my journal, but, I'm making good progress. Also, apologies for the repo upload-overlap. This is being written on the 14th of Feb--but I accidentally didn't commit until the night had passed on into the 15th.