03/03/2023, Updated FIles - FizzBuzz Part 1 Complete
Unfortunately, I folded. This did enable me to complete the first step of the fizzbuzz challenge, however, I am frustrated I wasn't able to figure it out on my own. Like I had mentioned previously, I managed to make the function work properly except for the fact that it wasn't repeatable unless I manually entered another line of code.

The solution ended up being the increment operator. I'm frustrated about this because this was the first thing I tried (and, in general, tried multiple times) and it hadn't worked due to syntax or misusage.

If I've learned something from this, it's that for coding, although there are many ways to achieve the same outcome--the most obvious answer will almost always be one of them. It's about having a theory and knowing how to implement it properly just as much as it is the actual coding.

I think I just lack experience, which is why from here on I'll be dedicating more time exclusively to practice and constructive feedback so I can improve my knowledge and my skill with JavaScript.