03/05/2023 - Updated Files, Final FizzBuzz Version
I finally finished the FIzzBuzz challenge, and the second part was much easier than the first. Here is the final version of the code. If there's something I've learned, it's that the order of your code matters!! Your code can be fine (and/or you have the right theory) but if your order is wrong it can completely wreck your code.